requirements.txt
langchain
langchain_community
openai
python-dotenv
faiss-cpu
tiktoken
flask
flask-cors
streamlit
langchain_groq